Refactor comparison into TestFlightProgramDifference
